+++ index.html
@@ -338,7 +338,7 @@
     </div>
 
 <script src="prompts.js"></script>
-<script src="src/main.js"></script>
+<script type="module" src="src/main.js"></script>
 </body>
 </html>
